+++ geotrek/trekking/filters.py
@@ -10,7 +10,7 @@ class TrekFilterSet(MapEntityFilterSet):
         model = Trek
         fields = ['published', 'difficulty', 'duration', 'themes', 'networks',
                   'practice', 'accessibilities', 'route', 'is_park_centered',
-                  'structure']
+                  'structure', 'source']
 
 
 class POITrekFilter(TopologyFilter):